requirements.txt
fastapi
uvicorn
pydantic
httpx
python-multipart
github
gitpython

